refactor: fix Addressability directory
